requirements.txt
gunicorn
flask
requests
pandas
cloudinary
matplotlib
anytree
imgkit
pyyaml